moved data structures to new file include/structs.h, building wall texture rendering
