+++ srcs/minimap/init_minimap.c
@@ -6,7 +6,7 @@
 /*   By: prynty <prynty@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/03 13:34:46 by prynty            #+#    #+#             */
-/*   Updated: 2025/03/06 14:13:38 by prynty           ###   ########.fr       */
+/*   Updated: 2025/03/10 16:55:31 by prynty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
