+++ srcs/minimap/minimap.c
@@ -11,7 +11,6 @@
 static void	update_movement_minimap_float(t_game *game, double x, double y)
 {
 	game->map.map[(int)game->player.pos_y][(int)game->player.pos_x] = '0';
-	// game->map.map[(int)y][(int)x] = 'N';
 	game->minimap.player->instances[0].x = (int)x * TILESIZE;
 	game->minimap.player->instances[0].y = (int)y * TILESIZE;
 }
