+++ srcs/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: prynty <prynty@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/25 15:16:34 by mrahmat-          #+#    #+#             */
-/*   Updated: 2025/03/10 14:42:19 by prynty           ###   ########.fr       */
+/*   Updated: 2025/03/10 17:16:07 by prynty           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,18 +30,29 @@ int	print_error(char *msg)
 	return (FALSE);
 }
 
-// void	game_hook(void *param)
-// {
-// 	t_game	*game;
+static void	key_move(t_game *game, double x, double y)
+{
+	if (mlx_is_key_down(game->mlx, MLX_KEY_W))
+		move_player_minimap(game, x + game->player.dir_x * MOVE_SPEED, y);
+	else if (mlx_is_key_down(game->mlx, MLX_KEY_S))
+		move_player_minimap(game, x - game->player.dir_x * MOVE_SPEED, y);
+	else if (mlx_is_key_down(game->mlx, MLX_KEY_A))
+		move_player_minimap(game, x, y - game->player.dir_y * MOVE_SPEED);
+	else if (mlx_is_key_down(game->mlx, MLX_KEY_D))
+		move_player_minimap(game, x, y + game->player.dir_y * MOVE_SPEED);
+}
 
-// 	game = (t_game *)param;
-// 	minimap(game);
-// }
+static void	key_rotate(t_game *game)
+{
+	if (mlx_is_key_down(game->mlx, MLX_KEY_LEFT))
+		rotate_player(game, false);
+	else if (mlx_is_key_down(game->mlx, MLX_KEY_RIGHT))
+		rotate_player(game, true);
+}
 
 void	key_hooks(mlx_key_data_t data, void *param)
 {
 	t_game	*game;
-	// mlx_key_data_t	data;
 	double	x;
 	double	y;
 
@@ -50,29 +61,10 @@ void	key_hooks(mlx_key_data_t data, void *param)
 	y = game->player.pos_y;
 	if (data.key == MLX_KEY_ESCAPE && data.action == MLX_RELEASE)
 		mlx_close_window(game->mlx);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_W))
-		move_player_minimap(game, x + MOVE_SPEED, y);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_S))
-		move_player_minimap(game, x - MOVE_SPEED, y);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_A))
-		move_player_minimap(game, x, y - MOVE_SPEED);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_D))
-		move_player_minimap(game, x, y + MOVE_SPEED);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_LEFT))
-		rotate_player(game, false);
-	else if (mlx_is_key_down(game->mlx, MLX_KEY_RIGHT))
-		rotate_player(game, true);
+	key_move(game, x, y);
+	key_rotate(game);
 	render_world(game);
 }
-
-// void	game_hook(void *param)
-// {
-// 	t_game	*game;
-	
-// 	game = (t_game *)param;
-// 	key_hooks(game);
-// 	// rotate_player(game, right);
-// }
 
 void	cleanup(t_game *game)
 {
